market basket anlasis phase 1
After checking the basic information of our dataset, we came up with the idea to implement market basket analysis with a focus on the tags. Each dataset row is treated as an individual transaction, combining 'Hotel_Name' with ‘Tags’ to represent items in a transaction. By applying the Apriori algorithm, we intend to identify frequent itemsets and generate association rules, using hotel names as antecedents to highlight the most commonly associated tags in the reviews pertaining to specific hotels.

We applied the market basket analysis and defined a function to let the customer input a hotel name and see the most frequently mentioned tags in order to see if the hotel is a good fit for their trip.