Market basket analysis 1.2
After checking the basic information of our dataset, we came up with the idea to implement market basket analysis with a focus on the tags. Each dataset row is treated as an individual transaction, combining 'Hotel_Name' with ‘Tags’ to represent items in a transaction. By applying the Apriori algorithm, we intend to identify frequent itemsets and generate association rules, using hotel names as antecedents to highlight the most commonly associated tags in the reviews pertaining to specific hotels.